requirements.txt
openai
pandas
langchain
langchain-community
langchain-experimental
python-dotenv
openpyxl
fastapi
uvicorn
psycopg2-binary
python-dotenv
llama-cpp-python
